fix: move Trace to components
