+++ web/src/components/Trace/TraceDiagram.tsx
