+++ web/src/pages/Test/Test.tsx
@@ -3,7 +3,7 @@ import Title from 'antd/lib/typography/Title';
 import {useParams} from 'react-router-dom';
 import {useGetTestByIdQuery} from '../../services/TestService';
 
-import Trace from '../Trace';
+import Trace from '../../components/Trace';
 import Assertions from './Assertions';
 import * as S from './Test.styled';
